Add int, string and byte message tests
